Clean up implementation and docs
